+++ event-store-management/src/main/java/uk/gov/justice/services/eventstore/management/catchup/events/CatchupCompletedForSubscriptionEvent.java
@@ -1,5 +1,6 @@
 package uk.gov.justice.services.eventstore.management.catchup.events;
 
+import uk.gov.justice.services.eventstore.management.catchup.commands.CatchupType;
 import uk.gov.justice.services.jmx.api.command.SystemCommand;
 
 import java.time.ZonedDateTime;
@@ -7,6 +8,7 @@ import java.util.Objects;
 
 public class CatchupCompletedForSubscriptionEvent {
 
+    private final CatchupType catchupType;
     private final String subscriptionName;
     private final String eventSourceName;
     private final String componentName;
@@ -15,18 +17,23 @@ public class CatchupCompletedForSubscriptionEvent {
     private final int totalNumberOfEvents;
 
     public CatchupCompletedForSubscriptionEvent(
-            final String subscriptionName,
+            final CatchupType catchupType, final String subscriptionName,
             final String eventSourceName,
             final String componentName,
             final SystemCommand target,
             final ZonedDateTime catchupCompletedAt,
             final int totalNumberOfEvents) {
+        this.catchupType = catchupType;
         this.subscriptionName = subscriptionName;
         this.eventSourceName = eventSourceName;
         this.componentName = componentName;
         this.target = target;
         this.catchupCompletedAt = catchupCompletedAt;
         this.totalNumberOfEvents = totalNumberOfEvents;
+    }
+
+    public CatchupType getCatchupType() {
+        return catchupType;
     }
 
     public String getSubscriptionName() {
@@ -59,6 +66,7 @@ public class CatchupCompletedForSubscriptionEvent {
         if (!(o instanceof CatchupCompletedForSubscriptionEvent)) return false;
         final CatchupCompletedForSubscriptionEvent that = (CatchupCompletedForSubscriptionEvent) o;
         return totalNumberOfEvents == that.totalNumberOfEvents &&
+                catchupType == that.catchupType &&
                 Objects.equals(subscriptionName, that.subscriptionName) &&
                 Objects.equals(eventSourceName, that.eventSourceName) &&
                 Objects.equals(componentName, that.componentName) &&
@@ -68,13 +76,14 @@ public class CatchupCompletedForSubscriptionEvent {
 
     @Override
     public int hashCode() {
-        return Objects.hash(subscriptionName, eventSourceName, componentName, target, catchupCompletedAt, totalNumberOfEvents);
+        return Objects.hash(catchupType, subscriptionName, eventSourceName, componentName, target, catchupCompletedAt, totalNumberOfEvents);
     }
 
     @Override
     public String toString() {
         return "CatchupCompletedForSubscriptionEvent{" +
-                "subscriptionName='" + subscriptionName + '\'' +
+                "catchupType=" + catchupType +
+                ", subscriptionName='" + subscriptionName + '\'' +
                 ", eventSourceName='" + eventSourceName + '\'' +
                 ", componentName='" + componentName + '\'' +
                 ", target=" + target +