+++ event-store-management/src/main/java/uk/gov/justice/services/eventstore/management/catchup/events/CatchupStartedForSubscriptionEvent.java
@@ -1,20 +1,31 @@
 package uk.gov.justice.services.eventstore.management.catchup.events;
 
+import uk.gov.justice.services.eventstore.management.catchup.commands.CatchupType;
+
 import java.time.ZonedDateTime;
 import java.util.Objects;
 
 public class CatchupStartedForSubscriptionEvent {
 
     private final String subscriptionName;
+    private final CatchupType catchupType;
     private final ZonedDateTime catchupStartedAt;
 
-    public CatchupStartedForSubscriptionEvent(final String subscriptionName, final ZonedDateTime catchupStartedAt) {
+    public CatchupStartedForSubscriptionEvent(
+            final String subscriptionName,
+            final CatchupType catchupType,
+            final ZonedDateTime catchupStartedAt) {
         this.subscriptionName = subscriptionName;
+        this.catchupType = catchupType;
         this.catchupStartedAt = catchupStartedAt;
     }
 
     public String getSubscriptionName() {
         return subscriptionName;
+    }
+
+    public CatchupType getCatchupType() {
+        return catchupType;
     }
 
     public ZonedDateTime getCatchupStartedAt() {
@@ -27,18 +38,20 @@ public class CatchupStartedForSubscriptionEvent {
         if (!(o instanceof CatchupStartedForSubscriptionEvent)) return false;
         final CatchupStartedForSubscriptionEvent that = (CatchupStartedForSubscriptionEvent) o;
         return Objects.equals(subscriptionName, that.subscriptionName) &&
+                catchupType == that.catchupType &&
                 Objects.equals(catchupStartedAt, that.catchupStartedAt);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(subscriptionName, catchupStartedAt);
+        return Objects.hash(subscriptionName, catchupType, catchupStartedAt);
     }
 
     @Override
     public String toString() {
         return "CatchupStartedForSubscriptionEvent{" +
-                "eventSourceName='" + subscriptionName + '\'' +
+                "subscriptionName='" + subscriptionName + '\'' +
+                ", catchupType=" + catchupType +
                 ", catchupStartedAt=" + catchupStartedAt +
                 '}';
     }