+++ event-store-management/src/main/java/uk/gov/justice/services/eventstore/management/catchup/process/CatchupsInProgressCache.java
@@ -1,5 +1,9 @@
 package uk.gov.justice.services.eventstore.management.catchup.process;
 
+import static uk.gov.justice.services.eventstore.management.catchup.commands.CatchupType.INDEX_CATCHUP;
+
+import uk.gov.justice.services.eventstore.management.catchup.commands.CatchupType;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
@@ -7,29 +11,38 @@ import java.util.concurrent.ConcurrentHashMap;
 
 public class CatchupsInProgressCache {
 
-    private final Map<String, CatchupInProgress> catchupsInProgress = new ConcurrentHashMap<>();
+    private final Map<String, CatchupInProgress> eventCatchupsInProgress = new ConcurrentHashMap<>();
+    private final Map<String, CatchupInProgress> indexCatchupsInProgress = new ConcurrentHashMap<>();
 
-    public void removeAll() {
-        catchupsInProgress.clear();
+    public void removeAll(final CatchupType catchupType) {
+        getCache(catchupType).clear();
     }
 
-    public void addCatchupInProgress(final CatchupInProgress catchupInProgress) {
-        catchupsInProgress.put(catchupInProgress.getSubscriptionName(), catchupInProgress);
+    public void addCatchupInProgress(final CatchupInProgress catchupInProgress, final CatchupType catchupType) {
+        getCache(catchupType).put(catchupInProgress.getSubscriptionName(), catchupInProgress);
     }
 
-    public CatchupInProgress removeCatchupInProgress(final String subscriptionName) {
-        return catchupsInProgress.remove(subscriptionName);
+    public CatchupInProgress removeCatchupInProgress(final String subscriptionName, final CatchupType catchupType) {
+        return getCache(catchupType).remove(subscriptionName);
     }
 
-    public boolean isCatchupInProgress(final String subscriptionName) {
-       return catchupsInProgress.containsKey(subscriptionName);
+    public boolean isCatchupInProgress(final String subscriptionName, final CatchupType catchupType) {
+       return getCache(catchupType).containsKey(subscriptionName);
     }
 
-    public List<CatchupInProgress> getAllCatchupsInProgress() {
-        return new ArrayList<>(catchupsInProgress.values());
+    public List<CatchupInProgress> getAllCatchupsInProgress(final CatchupType catchupType) {
+        return new ArrayList<>(getCache(catchupType).values());
     }
 
-    public boolean noCatchupsInProgress() {
-        return catchupsInProgress.isEmpty();
+    public boolean noCatchupsInProgress(final CatchupType catchupType) {
+        return getCache(catchupType).isEmpty();
+    }
+
+    private Map<String, CatchupInProgress> getCache(final CatchupType catchupType) {
+        if (catchupType == INDEX_CATCHUP) {
+            return indexCatchupsInProgress;
+        }
+
+        return eventCatchupsInProgress;
     }
 }