+++ event-store-management/src/test/java/uk/gov/justice/services/eventstore/management/catchup/process/EventCatchupProcessorTest.java
@@ -11,6 +11,7 @@ import static org.mockito.Mockito.inOrder;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
+import static uk.gov.justice.services.eventstore.management.catchup.commands.CatchupType.EVENT_CATCHUP;
 
 import uk.gov.justice.services.common.util.UtcClock;
 import uk.gov.justice.services.event.sourcing.subscription.catchup.consumer.manager.ConcurrentEventStreamConsumerManager;
@@ -19,7 +20,6 @@ import uk.gov.justice.services.eventsourcing.repository.jdbc.event.MissingEventN
 import uk.gov.justice.services.eventsourcing.repository.jdbc.event.PublishedEvent;
 import uk.gov.justice.services.eventsourcing.source.api.service.core.PublishedEventSource;
 import uk.gov.justice.services.eventstore.management.catchup.events.CatchupCompletedForSubscriptionEvent;
-import uk.gov.justice.services.eventstore.management.catchup.events.CatchupRequestedEvent;
 import uk.gov.justice.services.eventstore.management.catchup.events.CatchupStartedForSubscriptionEvent;
 import uk.gov.justice.services.jmx.api.command.SystemCommand;
 import uk.gov.justice.services.subscription.ProcessedEventTrackingService;
@@ -79,9 +79,13 @@ public class EventCatchupProcessorTest {
 
         final Subscription subscription = mock(Subscription.class);
         final PublishedEventSource publishedEventSource = mock(PublishedEventSource.class);
-        final CatchupRequestedEvent catchupRequestedEvent = mock(CatchupRequestedEvent.class);
-        final CatchupSubscriptionContext catchupSubscriptionContext = new CatchupSubscriptionContext(componentName, subscription, catchupRequestedEvent);
         final SystemCommand systemCommand = mock(SystemCommand.class);
+
+        final CatchupSubscriptionContext catchupSubscriptionContext = new CatchupSubscriptionContext(
+                componentName,
+                subscription,
+                EVENT_CATCHUP,
+                systemCommand);
 
         final PublishedEvent publishedEvent_1 = mock(PublishedEvent.class);
         final PublishedEvent publishedEvent_2 = mock(PublishedEvent.class);
@@ -102,7 +106,6 @@ public class EventCatchupProcessorTest {
         when(concurrentEventStreamConsumerManager.add(publishedEvent_1, subscriptionName)).thenReturn(1);
         when(concurrentEventStreamConsumerManager.add(publishedEvent_2, subscriptionName)).thenReturn(1);
         when(concurrentEventStreamConsumerManager.add(publishedEvent_3, subscriptionName)).thenReturn(1);
-        when(catchupRequestedEvent.getTarget()).thenReturn(systemCommand);
 
         eventCatchupProcessor.performEventCatchup(catchupSubscriptionContext);
 
@@ -113,6 +116,7 @@ public class EventCatchupProcessorTest {
 
         inOrder.verify(catchupStartedForSubscriptionEventFirer).fire(new CatchupStartedForSubscriptionEvent(
                 subscriptionName,
+                EVENT_CATCHUP,
                 catchupStartedAt));
 
         inOrder.verify(concurrentEventStreamConsumerManager).add(publishedEvent_1, subscriptionName);
@@ -121,6 +125,7 @@ public class EventCatchupProcessorTest {
         inOrder.verify(concurrentEventStreamConsumerManager).waitForCompletion();
 
         inOrder.verify(catchupCompletedForSubscriptionEventFirer).fire(new CatchupCompletedForSubscriptionEvent(
+                EVENT_CATCHUP,
                 subscriptionName,
                 eventSourceName,
                 componentName,
@@ -129,7 +134,7 @@ public class EventCatchupProcessorTest {
                 events.size()));
 
         verify(logger).info("Catching up from Event Number: " + eventNumberFrom);
-        verify(logger).info("Starting catch up for Event Number: " + (eventNumberFrom + 1L));
+        verify(logger).info("Event catch up for Event Number: " + (eventNumberFrom + 1L));
     }
 
     @Test
@@ -146,9 +151,13 @@ public class EventCatchupProcessorTest {
 
         final Subscription subscription = mock(Subscription.class);
         final PublishedEventSource publishedEventSource = mock(PublishedEventSource.class);
-        final CatchupRequestedEvent catchupRequestedEvent = mock(CatchupRequestedEvent.class);
-        final CatchupSubscriptionContext catchupSubscriptionContext = new CatchupSubscriptionContext(componentName, subscription, catchupRequestedEvent);
         final SystemCommand systemCommand = mock(SystemCommand.class);
+
+        final CatchupSubscriptionContext catchupSubscriptionContext = new CatchupSubscriptionContext(
+                componentName,
+                subscription,
+                EVENT_CATCHUP,
+                systemCommand);
 
         final PublishedEvent publishedEvent_1 = mock(PublishedEvent.class);
         final PublishedEvent publishedEvent_2 = mock(PublishedEvent.class);
@@ -170,7 +179,6 @@ public class EventCatchupProcessorTest {
         when(concurrentEventStreamConsumerManager.add(publishedEvent_1, subscriptionName)).thenReturn(1);
         when(concurrentEventStreamConsumerManager.add(publishedEvent_2, subscriptionName)).thenReturn(1);
         when(concurrentEventStreamConsumerManager.add(publishedEvent_3, subscriptionName)).thenReturn(1);
-        when(catchupRequestedEvent.getTarget()).thenReturn(systemCommand);
 
         try {
             eventCatchupProcessor.performEventCatchup(catchupSubscriptionContext);