+++ event-store-management/src/test/java/uk/gov/justice/services/eventstore/management/catchup/process/EventCatchupRunnerTest.java
@@ -4,10 +4,11 @@ import static java.util.Arrays.asList;
 import static org.mockito.Mockito.inOrder;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
+import static uk.gov.justice.services.eventstore.management.catchup.commands.CatchupType.EVENT_CATCHUP;
 
 import uk.gov.justice.services.common.util.UtcClock;
-import uk.gov.justice.services.eventstore.management.catchup.events.CatchupRequestedEvent;
 import uk.gov.justice.services.eventstore.management.catchup.events.CatchupStartedEvent;
+import uk.gov.justice.services.jmx.api.command.CatchupCommand;
 import uk.gov.justice.subscription.domain.subscriptiondescriptor.SubscriptionsDescriptor;
 import uk.gov.justice.subscription.registry.SubscriptionsDescriptorsRegistry;
 
@@ -53,12 +54,12 @@ public class EventCatchupRunnerTest {
         final SubscriptionsDescriptor subscriptionsDescriptor_1 = mock(SubscriptionsDescriptor.class);
         final SubscriptionsDescriptor subscriptionsDescriptor_2 = mock(SubscriptionsDescriptor.class);
 
-        final CatchupRequestedEvent catchupRequestedEvent = mock(CatchupRequestedEvent.class);
+        final CatchupCommand catchupCommand = new CatchupCommand();
 
         when(clock.now()).thenReturn(startTime, endTime);
         when(subscriptionsDescriptorsRegistry.getAll()).thenReturn(asList(subscriptionsDescriptor_1, subscriptionsDescriptor_2));
 
-        eventCatchupRunner.runEventCatchup(catchupRequestedEvent);
+        eventCatchupRunner.runEventCatchup(EVENT_CATCHUP, catchupCommand);
 
         final InOrder inOrder = inOrder(
                 logger,
@@ -66,9 +67,9 @@ public class EventCatchupRunnerTest {
                 eventCatchupByComponentRunner
         );
 
-        inOrder.verify(logger).info("Received CatchupRequestedEvent");
-        inOrder.verify(catchupStartedEventFirer).fire(new CatchupStartedEvent(startTime));
-        inOrder.verify(eventCatchupByComponentRunner).runEventCatchupForComponent(subscriptionsDescriptor_1, catchupRequestedEvent);
-        inOrder.verify(eventCatchupByComponentRunner).runEventCatchupForComponent(subscriptionsDescriptor_2, catchupRequestedEvent);
+        inOrder.verify(logger).info("Received CatchupRequestedEvent for EVENT_CATCHUP");
+        inOrder.verify(catchupStartedEventFirer).fire(new CatchupStartedEvent(EVENT_CATCHUP, startTime));
+        inOrder.verify(eventCatchupByComponentRunner).runEventCatchupForComponent(subscriptionsDescriptor_1, EVENT_CATCHUP, catchupCommand);
+        inOrder.verify(eventCatchupByComponentRunner).runEventCatchupForComponent(subscriptionsDescriptor_2, EVENT_CATCHUP, catchupCommand);
     }
 }