+++ CHANGELOG.md
@@ -7,6 +7,7 @@ on [Keep a CHANGELOG](http://keepachangelog.com/). This project adheres to
 
 ### Changed
 - Subscriptions are no longer run asynchronously during catchup
+- Event catchup and Indexer catchup now run the same code
 
 ## [2.0.22] - 2019-09-24
 ### Changed